Core - added build support for KEYP8 section. Reworked section build logic to immediately add the section rather than as a list at once, so that sections can be built that rely on other sections.

+++ XUIHelper.Core/XUIHelper.Core/XUR/Sections/IKEYPSection.cs
@@ -10,6 +10,6 @@ namespace XUIHelper.Core
     {
         const int ExpectedMagic = 0x4B455950;
 
-        List<ulong> PropertyIndexes { get; }
+        List<uint> PropertyIndexes { get; }
     }
 }
